Remove singletons
The template loader was a singleton in the Response class, that was completelly
removed. The template loader is initialized in the make_app() function and
stored along with the application instance.

Rendering templates is now done via a method in a custom Request class (see
vezilka/lib/utils.py). The views are configured to use the new Request class by
default (vezilka/views.py).